mitmproxy/rewrite_to_local/requirements.txt
asgiref==3.5.2
Brotli==1.0.9
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
cryptography==38.0.4
Flask==2.2.5
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==3.0.2
mitmproxy==9.0.1
mitmproxy_wireguard==0.1.23
msgpack==1.0.8
passlib==1.7.4
protobuf==4.25.8
publicsuffix2==2.20191221
pyasn1==0.6.1
pycparser==2.22
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
python-dotenv==1.0.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
sortedcontainers==2.4.0
tornado==6.5.1
urwid==2.1.2
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.19.0
